Update: add CHROME_BIN Puppeteer seam for CI tests
When executing development tests, inspect CHROME_BIN and use that
executable if defined. This is necessary for the execution of CI tests
which use the Debian system install Chromium package, not the NPM
Puppeteer installed version. When this variable is undefined, the
Puppeteer packaged version is used.

For production, service-runner supports an environment variable syntax,
so the prod config has been updated to use CHROME_BIN which will be
defined in the package.json file in
I48ba251f6b89c82ba69e0a187c45554e5f73f64f. Unfortunately, this syntax
cannot be used for development since the test runner loads and adjusts
the YAML config before invoking it so no parsing is done and
`'{env(CHROME_BIN)}'` is interpreted as a string.

After this patch, `CHROME_BIN=/bin/false npm t` should fail and `npm t`
should still succeed.

https://github.com/wikimedia/service-runner/

Bug: T179552
Change-Id: I259c22555738d60b5b10f5e00dedd377417b0ea1