Check the article existence before adding it to the queue
Before we enqueue the job we should check for article existence.
For example the service-checker will try to render a non-existing
article every minute, there is no need to queue that job and
waste space in the queue.

Before adding task to the queue, lets check the article html
using restbase page/html/{TITLE} endpoint. That endpoint will
return 404 if article doesn't exist.

Bug: T197856
Change-Id: I3fd37da6ff8599a778fc7d7ba483678b6b60ddc5