Update Puppeteer to v1.7.0
A part of chore wheel - to keep the Puppetteer version up to date,
the latest version is 1.7.0, we're jumping from v.1.5.0.

The latest 1.7.0 version brings the `puppeteer-core` package, that
by default do not download chromium. Thanks to that we can drop
`PUPPETEER_SKIP_CHROMIUM_DOWNLOAD` env variable as it is not
necessary. Most problably we can also skip the `CHROME_BIN` variable
as the path to chromium executable has to be always hardcoded in
the config file from now on.

Changes introduced in puppeteer that are related to proton:
 - in 1.7.0
   - page.pdf() accepts a new argument preferCSSPageSize (12e3510)
   - introduced puppeteer-core - that doesn't download chromium,
   we don't need to handle SKIP_CHROMIUM_DOWNLOAD any more
 -  in 1.6.1
   - #2927 - Bad font rendering on Linux (although it doesn't
   affect our instance)
 - in 1.6.0
   - #2685 - Page.pdf - Colors rendering

Change-Id: I10e8308149ef8067a3830cdd0f2ac539edeee7a1