Handle the stopping chromium rendering correctly
The previous implementation struggled with handling abort requests.
In simple way, there was no possibility to stop the started task.
Script was able to close the browser, but the started rendering job
had to fail itself in any way. Usually it was couple seconds
before Puppeteer threw `Protol error: Target close` to even up to
seconds with `Navigation Timeout`. After failure we had to handle
the events bubbling, and fact that sometimes `articleToPDF` promise
resolved with an undefined `pdf` variable.
Also, even puppeteer wasn't processing any data, it was still keeping
the spot in the running jobs queue and preventing other jobs to be
picked up.

Current implementation defines it's own promise (is not based on the
puppeteer promise as it was before) because there was no way to
reject the promise when job was aborted. Puppeteer listens when
browser gets closed (by litening to `disconnect` event) and rejects
promise immediately which causes to free the spot in the queue and
ignores any other crazy/unexpected behaviour.

Changes:
 - change the queue.callbackkErrors as someone pointed there might
 be problem with handling queueBusy (see removed FIX: from
 html2pdf-v1.js)
 - added handling the `abort` exception which doesn't cause
 anything to be logged, nor any stats to incremented.
 - added important comments about the code flow (queue handling
 and the fact that Puppeteer likes to throw exceptions when browser
 is closed and PDF rendering is still in progress)
 - articleToPDF uses the BBPromise now, not the puppeteer promise
 - closeBrowser returns a BBPromise which resolves when browser
 is closed or killed
 - puppeteer will stop the render job when browser is closed

Bug: T181623
Change-Id: Iaf6ea75d596c52fc99c70467902d1e11999900d6