Handle undefined response
After enabling 25% of traffic to the proton service, service
started to log `Cannot read property 'ok' of null`. This is
clearly the case when we get `null` instead of Puppeteer
response. There might be many reasons, lets log when that
specific case happens.

After we gather some logs, we will need to follow up, and find
why do we get an undefined response from Puppeteer.
Is it a puppeteer issue? Does it happen for not-aborted renders?

Bug: T186748
Change-Id: I3d74e2a8dccf465a41e740799f0bc929d6ce8e66