Rewrite Queue to promise-way flow
All parts of code uses promises but the Queue.js was the only
file that uses AsyncQueue which makes the flow inconsistent in
the Proton renderer.

After research I couldn't find a queue library that would fulfill
our requirements:
 - set up render concurrency and max queue size
 - allow timeouts both for in queue and in progress states
 - allow to cancel the queued task
Because of that I had to come-up with bespoke queue implementation
that fulfills our requirements.I tried to make the Queue and
QueueItem as generic as possible. I'd like to turn it into a
library some time in the future. The Queue uses EventEmitter
to publish state changes, route attaches listeners to different
states and logs/update metrics accordingly.
Queue.push() can take just the promise builder and create the
QueueItem by itself.

Additional changes:
 - removed tracking queue errors as those are properly represented
 by the endpoint http response codes
 - removed Queue health as metrics provide much better data
 - removed async dependency as it is not required any more

Bug: T204055
Change-Id: I7698cddcb2135cdfd8fe9cd09cb752b069880a71