Update puppeteer to latest 1.9.0
Introduced in 1.9.0
 - no changes that affect our library

Introduced in 1.8.0:
 - Drop the CHROME_BIN variable as it is not necessary,
Puppeteer provides it's own PUPPETEER_EXECUTABLE_PATH. It's better
to use built-in puppeteer features instead of coming with our own
solution (which is cleary a duplicated feature).

Introduced in 1.7.0:
 - The 1.7.0 version brings the `puppeteer-core` package, that
by default do not download chromium. Thanks to that we can drop
`PUPPETEER_SKIP_CHROMIUM_DOWNLOAD` env variable as it is not
necessary. Most problably we can also skip the `CHROME_BIN` variable
as the path to chromium executable has to be always hardcoded in
the config file from now on.
 - page.pdf() accepts a new argument preferCSSPageSize (12e3510)

Introduced in 1.6.1
   - #2927 - Bad font rendering on Linux (although it doesn't
   affect our instance)
 - in 1.6.0
   - #2685 - Page.pdf - Colors rendering

Bug: T207416
Change-Id: Ib94d7bbdf1a3edf6f8c0fecfc4c69b3625109a96