Hygiene: Don't call restbase when queue is full
This task is a part of hygiene fixes (no changes to main rendering
logic) before project handoff.

If the queue is full, there is no need to call restbase to verify
article existance. This is mostly a minor optimalization.
Previously, the system first would check for the article existence
by calling restbase `/page/title/` endpoing and then return queue
full error. If we can avoid calling restbase, and immediately return
queue full we could save some resources and decrease network traffic.
It's helpful in cases when someone tries to ddois the service.

Change-Id: I29a7a34a94769548c78d1d9795dd6e440ecfe2f4