Hygiene: Stop requiring APP_ENABLE_CANCELLABLE_PROMISES
The service currently attempts to configure Bluebird promises to be
cancellable by way of service-runner, via the
APP_ENABLE_CANCELLABLE_PROMISES environment variable. At least in my
local environment, despite being explicitly required, this doesn't
really seem to work for some reason; it also clutters up the code;
also, it's totally unnecessary, because we're already directly
configuring promises to be cancellable in /server.js.

This patch drops the apparently-pointless requirement of using of
using the the environment variable.

Change-Id: Ibd928c212ca1f3b6ae8b77a7227a52eb4dad1b84