Run chromium with --single-process and --no-zygote flags
Attempts to prevent leaking zombie chromium processes by running
chromium with the --single-process and --no-zygote flags.

Flag descriptions:
https://kapeli.com/cheat_sheets/Chromium_Command_Line_Switches.docset/Contents/Resources/Documents/index

GitHub issue where this approach was reported to be successful:
https://github.com/puppeteer/puppeteer/issues/1825

Chromium related docs (for background):
https://www.chromium.org/developers/design-documents/multi-process-architecture
https://chromium.googlesource.com/chromium/src.git/+/master/docs/linux/zygote.md

N.B. If the approach is agreed upon and this patch merged, the
change will also need to be made separately for production in
operations/deployment-charts.

Bug: T257679
Change-Id: I8420c4b9e967e098f279fc26e0128e6169a204c6