Improve catchall error logging
An occasional error appears in the logs with the message, "Unexpected
error: [Object object]." This updates the logged message to use the
`message` property of the passed-in error if it's an object. (To get
the line length under 100 characters, I removed the redundant word
"unexpected," since errors are always unexpected.)

Hygiene: Fixed a JSDoc type annotation that was almost certainly never
correct.

Change-Id: I5861e9440f119d54cdb668642296ad8459747386
Hygiene: Minor grammatical fix to a comment.