Make PDF format optional and use a4 as default
RESTBase would default to the A4 format if not format is specified
and Proton should behave the same way if no format is specified.
This is required to expose proton via RestGateway as majority
of callers do not pass format when requesting a PDF.

Log metrics for the "default" case (when no format is given
and Proton has to resolve to a default format from the config).

Co-Authored-by: D3r1ck01 <alangiderick@gmail.com>
Bug: T340630
Change-Id: Ica151e3cfc0dd3fbdcb69c5dfe06abb22e74e037